backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
httpx>=0.27.2
anthropic>=0.40.0
python-multipart>=0.0.9
pydantic>=2.11.0
aiofiles>=24.1.0
psycopg2-binary>=2.9.9
bcrypt>=4.2.0
PyJWT>=2.9.0
python-magic>=0.4.27
python-dotenv>=1.0.0
openpyxl>=3.1.0
apscheduler>=3.10.0
cryptography>=42.0.0
playwright>=1.40.0
# Super Agent 의존성
litellm>=1.30.0
python-docx>=1.1.0
python-pptx>=0.6.23
pdfplumber>=0.11.0
openai>=1.40.0
google-generativeai>=0.8.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
python-calamine>=0.2.0
